+++ Products/GenericSetup/browser/components.py
@@ -11,14 +11,15 @@
 #
 ##############################################################################
 """Components setup view.
-
-$Id$
 """
 
+import os.path
+
+import zope.formlib
+from Products.Five.browser.decode import processInputs
+from Products.Five.browser.decode import setPageEncoding
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from Products.Five.component.interfaces import IObjectManagerSite
-
-from five.formlib.formbase import PageEditForm
 from zope.component import adapts
 from zope.component import getMultiAdapter
 from zope.formlib import form
@@ -28,6 +29,9 @@ from zope.schema import Text
 
 from Products.GenericSetup.context import SetupEnviron
 from Products.GenericSetup.interfaces import IBody
+
+_FORMLIB_DIR = os.path.dirname(zope.formlib.__file__)
+_PAGEFORM_PATH = os.path.join(_FORMLIB_DIR, 'pageform.pt')
 
 
 class IComponentsSetupSchema(Interface):
@@ -58,14 +62,21 @@ class ComponentsSetupSchemaAdapter(object):
     body = property(_getBody, _setBody)
 
 
-class ComponentsSetupView(PageEditForm):
+class ComponentsSetupView(form.PageEditForm):
 
     """Components setup view for IObjectManagerSite.
     """
 
+    template = ViewPageTemplateFile(_PAGEFORM_PATH)
+
     label = u'Component Registry: XML Configuration'
 
     form_fields = form.FormFields(IComponentsSetupSchema)
+
+    def update(self):
+        processInputs(self.request)
+        setPageEncoding(self.request)
+        super(ComponentsSetupView, self).update()
 
     def setUpWidgets(self, ignore_request=False):
         super(ComponentsSetupView,
@@ -78,7 +89,7 @@ class ComponentsSetupTab(ComponentsSetupView):
     """Components setup ZMI tab for IObjectManagerSite.
     """
 
-    base_template = PageEditForm.template
+    base_template = ComponentsSetupView.template
 
     template = ViewPageTemplateFile('components.pt')